+++ Validator/Constraints/PercentRange.php
@@ -1,0 +1,16 @@
+<?php
+
+namespace NetTeam\Bundle\DDDBundle\Validator\Constraints;
+
+use Symfony\Component\Validator\Constraints\Range;
+
+/**
+ * Constraint dla PercentValidator
+ *
+ * @author Krzysztof Menżyk <krzysztof.menzyk@netteam.pl>
+ *
+ * @Annotation
+ */
+class PercentRange extends Range
+{
+}